refactor: split comb expression evaluator
